feat: remove maps
