--- src/Components/Maps.tsx
@@ -1,141 +1,0 @@
-import {
-  useLoadScript,
-  GoogleMap,
-  MarkerF,
-  CircleF,
-} from '@react-google-maps/api';
-import type { NextPage } from 'next';
-import { useMemo, useState } from 'react';
-import usePlacesAutocomplete, {
-  getGeocode,
-  getLatLng,
-} from 'use-places-autocomplete';
-import styles from '../styles/Home.module.css';
-
-const Maps: NextPage = () => {
-
-  // get the current location
-  const [lat, setLat] = useState<number>(0);
-  const [lng, setLng] = useState<number>(0);
-
-  // get the current location on first render
-  useMemo(() => {
-    if (navigator.geolocation) {
-      navigator.geolocation.getCurrentPosition((position) => {
-        setLat(position.coords.latitude);
-        setLng(position.coords.longitude);
-      });
-    }
-  }, []);
-
-  const libraries = useMemo(() => ['places'], []);
-  const mapCenter = useMemo(() => ({ lat: lat, lng: lng }), [lat, lng]);
-
-  const mapOptions = useMemo<google.maps.MapOptions>(
-    () => ({
-      disableDefaultUI: true,
-      clickableIcons: true,
-      scrollwheel: false,
-    }),
-    []
-  );
-
-  const { isLoaded } = useLoadScript({
-    googleMapsApiKey: process.env.NEXT_PUBLIC_GOOGLE_API_KEY as string,
-    libraries: libraries as any,
-  });
-
-  if (!isLoaded) {
-    return <p>Loading...</p>;
-  }
-
-  return (
-    <div className={styles.homeWrapper}>
-      <div className={styles.sidebar}>
-        <span>Shops</span>
-        {/* render Places Auto Complete and pass custom handler which updates the state */}
-        <PlacesAutocomplete
-          onAddressSelect={(address) => {
-            getGeocode({ address: address }).then((results) => {
-              const { lat, lng } = getLatLng(results[0]);
-              setLat(lat);
-              setLng(lng);
-            });
-          }}
-        />
-      </div>
-      <GoogleMap
-        options={mapOptions}
-        zoom={14}
-        center={mapCenter}
-        mapTypeId={google.maps.MapTypeId.ROADMAP}
-        mapContainerStyle={{ width: '800px', height: '800px' }}
-        onLoad={(map) => console.log('Map Loaded')}
-      >
-        <MarkerF
-          position={mapCenter}
-          onLoad={() => console.log('Marker Loaded')}
-        />
-      </GoogleMap>
-    </div>
-  );
-};
-
-const PlacesAutocomplete = ({
-  onAddressSelect,
-}: {
-  onAddressSelect?: (address: string) => void;
-}) => {
-  const {
-    ready,
-    value,
-    suggestions: { status, data },
-    setValue,
-    clearSuggestions,
-  } = usePlacesAutocomplete({
-    requestOptions: { componentRestrictions: { country: '' } },
-    debounce: 300,
-    cache: 86400,
-  });
-
-  const renderSuggestions = () => {
-    return data.map((suggestion) => {
-      const {
-        place_id,
-        structured_formatting: { main_text, secondary_text },
-        description,
-      } = suggestion;
-
-      return (
-        <li
-          key={place_id}
-          onClick={() => {
-            setValue(description, false);
-            clearSuggestions();
-            onAddressSelect && onAddressSelect(description);
-          }}
-        >
-          <strong>{main_text}</strong> <small>{secondary_text}</small>
-        </li>
-      );
-    });
-  };
-
-  return (
-    <div className={styles.autocompleteWrapper}>
-      <input
-        value={value}
-        className={styles.autocompleteInput}
-        disabled={!ready}
-        onChange={(e) => setValue(e.target.value)}
-        placeholder="123 Stariway To Heaven"
-      />
-
-      {status === 'OK' && (
-        <ul className={styles.suggestionWrapper}>{renderSuggestions()}</ul>
-      )}
-    </div>
-  );
-};
-
-export default Maps;
