+++ src/Components/form/ShopForm.tsx
@@ -1,9 +1,19 @@
-import Maps from '../Maps'
-import React, { useState, type ChangeEvent } from 'react'
+import React, { useMemo, useState, type ChangeEvent } from 'react'
 import type Shop from '../../Api/Models/Shop'
 import { ShopService } from '../../Api/Services'
+import {
+  useLoadScript,
+
+} from '@react-google-maps/api';
+import usePlacesAutocomplete, {
+  getGeocode,
+} from 'use-places-autocomplete';
+
+
+
 
 const ShopForm: React.FC = () => {
+  
   const initialShopState: Shop = {
     companyName: '',
     address: '',
@@ -15,7 +25,21 @@ const ShopForm: React.FC = () => {
     lat: '',
     long: ''
   }
+  
   const [shop, setShop] = useState<Shop>(initialShopState)
+
+  const libraries = useMemo(() => ['places'], []);
+
+  const { isLoaded } = useLoadScript({
+    googleMapsApiKey: process.env.NEXT_PUBLIC_GOOGLE_API_KEY as string,
+    libraries: libraries as any,
+  });
+
+  if (!isLoaded) {
+    return <p>Loading...</p>;
+  }
+    
+
 
   const handleInputChange = (event: ChangeEvent<HTMLInputElement>): void => {
     const { name, value } = event.target
@@ -45,8 +69,8 @@ const ShopForm: React.FC = () => {
           zipCode: response.data.zipCode,
           phone: response.data.phone,
           email: response.data.email,
-          lat: response.data.lat,
-          long: response.data.long
+          lat: Maps.lat,
+          long: Maps.lng
         })
         console.log(response.data)
       })
@@ -73,15 +97,15 @@ const ShopForm: React.FC = () => {
 
         <div className="form-group">
           <label htmlFor="address">Address</label>
-          <input
-            type="text"
-            className="form-control"
-            id="address"
-            required
-            value={shop.address}
-            onChange={handleInputChange}
-            name="address"
-          />
+          <PlacesAutocomplete
+          onAddressSelect={(address) => {
+            getGeocode({ address: address }).then((results) => {
+              const lat = () => results[0].geometry.location.lat().toString();
+              const lng = () => results[0].geometry.location.lng().toString();
+              setShop({ ...shop, lat: lat(), long: lng(), address: address });
+            });
+          }}
+        />
         </div>
 
         <div className="form-group">
@@ -148,7 +172,6 @@ const ShopForm: React.FC = () => {
             name="email"
           />
         </div>
-        <Maps />
         <button onClick={saveShop} className="btn btn-success">
           Submit
         </button>
@@ -156,5 +179,62 @@ const ShopForm: React.FC = () => {
     </div>
   )
 }
+
+const PlacesAutocomplete = ({
+  onAddressSelect,
+}: {
+  onAddressSelect?: (address: string) => void;
+}) => {
+  const {
+    ready,
+    value,
+    suggestions: { status, data },
+    setValue,
+    clearSuggestions,
+  } = usePlacesAutocomplete({
+    requestOptions: { componentRestrictions: { country: '' } },
+    debounce: 300,
+    cache: 86400,
+  });
+
+  const renderSuggestions = () => {
+    return data.map((suggestion) => {
+      const {
+        place_id,
+        structured_formatting: { main_text, secondary_text },
+        description,
+      } = suggestion;
+
+      return (
+        <li
+          key={place_id}
+          onClick={() => {
+            setValue(description, false);
+            clearSuggestions();
+            onAddressSelect && onAddressSelect(description);
+          }}
+        >
+          <strong>{main_text}</strong> <small>{secondary_text}</small>
+        </li>
+      );
+    });
+  };
+
+  return (
+    <div className="">
+      <input
+        value={value}
+        className=""
+        disabled={!ready}
+        onChange={(e) => setValue(e.target.value)}
+        placeholder="123 Stariway To Heaven"
+      />
+
+      {status === 'OK' && (
+        <ul className="">{renderSuggestions()}</ul>
+      )}
+    </div>
+  );
+};
 
 export default ShopForm
